+++ src/lepl/__init__.py
@@ -285,7 +285,7 @@ __all__ = [
         'constant_indent'
        ]
 
-__version__ = '4.3.1'
+__version__ = '4.3.2'
 
 if __version__.find('b') > -1:
     from logging import getLogger, basicConfig, WARN