+++ src/lepl/_test/__init__.py
@@ -67,11 +67,11 @@ def all():
     print('\n\n\n----------------------------------------------------------'
           '------------\n')
     if version[0] == '2':
-        print('Expect 4-5 failures + 2 errors in Python 2.6: {0:d}, {1:d} '
+        print('Expect 3-5 failures + 2 errors in Python 2.6: {0:d}, {1:d} '
               '(lenient comparison, format variation from address size, '
               'unicode ranges, weird string difference; json unicode)'
               .format(len(result.failures), len(result.errors)))
-        assert 4 <= len(result.failures) <= 5, len(result.failures)
+        assert 3 <= len(result.failures) <= 5, len(result.failures)
         assert 2 <= len(result.errors) <= 2, len(result.errors)
         target = TOTAL - 22 - 9 # no bin/cairo tests (22)
     else: